--- app/src/main/java/com/example/mindmoving/dao/PerfilCalibracionDAO.kt
@@ -1,2 +1,0 @@
-package com.example.mindmoving.dao
-
