--- app/src/main/java/com/example/mindmoving/dao/SesionEegDao.kt
@@ -1,2 +1,0 @@
-package com.example.mindmoving.dao
-
