--- app/src/main/java/com/example/mindmoving/dao/UsuarioDao.kt
@@ -1,18 +1,0 @@
-package com.example.mindmoving.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-import com.example.mindmoving.models.UsuarioEntity
-
-@Dao
-interface UsuarioDao {
-    @Insert
-    suspend fun insertar(usuario: UsuarioEntity)
-
-    @Query("SELECT * FROM usuarios WHERE email = :email LIMIT 1")
-    suspend fun buscarPorEmail(email: String): UsuarioEntity?
-
-    @Query("SELECT * FROM usuarios WHERE id = :id")
-    suspend fun obtenerPorId(id: Int): UsuarioEntity
-}
