+++ app/src/main/java/com/example/mindmoving/deprecatedTemporalmente/dao/SesionEegDao.kt
@@ -1,0 +1,2 @@
+package com.example.mindmoving.deprecatedTemporalmente.dao
+
