--- app/src/main/java/com/example/mindmoving/models/UsuarioEntity.kt
@@ -1,15 +1,0 @@
-package com.example.mindmoving.models
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "usuarios")
-data class UsuarioEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val nombre: String,
-    val email: String,
-    val passwordHash: String,
-    val fechaCreacion: Long = System.currentTimeMillis(),
-    val ultimoLoginExitoso: Long? = null,
-    val rol: String = "usuario"
-)
