+++ app/src/main/java/com/example/mindmoving/retrofit/ApiClient.kt
@@ -4,10 +4,17 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 object ApiClient {
-    private val retrofit = Retrofit.Builder()
-        .baseUrl("http://10.0.2.2:3000") // Emulador o tu IP real si es dispositivo físico
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
 
-    val apiService: ApiService = retrofit.create(ApiService::class.java)
+    private const val BASE_URL = "http://puto:3000/" //IP de tu pc en puto
+
+    private val retrofit: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+    fun getApiService(): ApiService {
+        return retrofit.create(ApiService::class.java)
+    }
 }
