+++ app/src/main/java/com/example/mindmoving/retrofit/ApiService.kt
@@ -1,20 +1,20 @@
 package com.example.mindmoving.retrofit
 
-import com.example.mindmoving.retrofit.modelsDate.GenericResponse
-import com.example.mindmoving.retrofit.modelsDate.LoginRequest
-import com.example.mindmoving.retrofit.modelsDate.LoginResponse
-import com.example.mindmoving.retrofit.modelsDate.RegisterRequest
-import com.example.mindmoving.retrofit.modelsDate.SesionEEGRequest
+import com.example.mindmoving.retrofit.models.GenericResponse
+import com.example.mindmoving.retrofit.models.LoginRequest
+import com.example.mindmoving.retrofit.models.LoginResponse
+import com.example.mindmoving.retrofit.models.RegisterRequest
+import com.example.mindmoving.retrofit.models.SesionEEGRequest
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
 interface ApiService {
     @POST("/api/login")
-    suspend fun login(@Body request: LoginRequest): Response<LoginResponse>
+    suspend fun loginUser(@Body request: LoginRequest): Response<LoginResponse>
 
-    @POST("/api/register")
-    suspend fun register(@Body request: RegisterRequest): Response<GenericResponse>
+    @POST("api/register")
+    suspend fun registerUser(@Body request: RegisterRequest): Response<GenericResponse>
 
     @POST("/api/sesiones")
     suspend fun guardarSesion(@Body sesion: SesionEEGRequest): Response<GenericResponse>
