+++ app/src/main/java/com/example/mindmoving/retrofit/models/GenericResponse.kt
@@ -1,0 +1,6 @@
+package com.example.mindmoving.retrofit.models
+
+data class GenericResponse(
+    val success: Boolean,
+    val message: String
+)
