+++ app/src/main/java/com/example/mindmoving/retrofit/models/LoginRequest.kt
@@ -1,0 +1,5 @@
+package com.example.mindmoving.retrofit.models
+
+data class LoginRequest(
+    val email: String,
+    val password: String)
