+++ app/src/main/java/com/example/mindmoving/retrofit/models/LoginResponse.kt
@@ -1,0 +1,3 @@
+package com.example.mindmoving.retrofit.models
+
+data class LoginResponse(val message: String, val userId: String)
