+++ app/src/main/java/com/example/mindmoving/retrofit/models/RegisterRequest.kt
@@ -1,0 +1,7 @@
+package com.example.mindmoving.retrofit.models
+
+data class RegisterRequest(
+    val email: String,
+    val password: String,
+    val username: String
+)
