+++ app/src/main/java/com/example/mindmoving/retrofit/models/SesionEEGRequest.kt
@@ -1,0 +1,11 @@
+package com.example.mindmoving.retrofit.models
+
+data class SesionEEGRequest(
+    val usuarioId: String,
+    val fechaHora: String, // formato ISO-8601
+    val duracion: Int,
+    val valorMedioAtencion: Float,
+    val valorMedioRelajacion: Float,
+    val valorMedioPestaneo: Float,
+    val comandosEjecutados: String
+)
