--- app/src/main/java/com/example/mindmoving/retrofit/modelsDate/GenericResponse.kt
@@ -1,5 +1,0 @@
-package com.example.mindmoving.retrofit.modelsDate
-
-data class GenericResponse(
-    val message: String
-)
