--- app/src/main/java/com/example/mindmoving/retrofit/modelsDate/LoginRequest.kt
@@ -1,5 +1,0 @@
-package com.example.mindmoving.retrofit.modelsDate
-
-data class LoginRequest(
-    val email: String,
-    val password: String)
