--- app/src/main/java/com/example/mindmoving/retrofit/modelsDate/LoginResponse.kt
@@ -1,3 +1,0 @@
-package com.example.mindmoving.retrofit.modelsDate
-
-data class LoginResponse(val message: String, val userId: String)
