--- app/src/main/java/com/example/mindmoving/retrofit/modelsDate/RegisterRequest.kt
@@ -1,3 +1,0 @@
-package com.example.mindmoving.retrofit.modelsDate
-
-data class RegisterRequest(val username: String, val email: String, val password: String)
