+++ app/src/main/res/xml/network_security_config.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<network-security-config>
+    <domain-config cleartextTrafficPermitted="true">
+        <domain includeSubdomains="false">puto</domain> <!-- Usa tu IP real del pc inalambrico  -->
+    </domain-config>
+</network-security-config>
